Send icon data using NRGBA colour-space
TRELLO: https://trello.com/c/Ekz2pwg6/2345-icons-look-funky

Under the hood `org.freedesktop.StatusNotifierItem.IconPixmap` expects
to receive an `ARRAY(INT, INT, ARRAY BYTE)` where the arguments are
width, height and a `ARGB32` representation of the image[^1]. The
`ARGB32` seems to be a serialized version of a `QPixmap`[^2].

At no point does the documentation specify whether the `ARGB` /
`QPixmap` pixels must use premultiplied-alpha colour or not. But in
testing it does appear to assume non-alpha-premultiplied colours. Also
interesting is QT specifies NARGB as `Format_ARGB32` and ARGB as
`Format_ARGB32_Premultiplied`. Which seems to imply non-premultiplied is
the default, despite being less common.

Usage of `for ... range` bumps minimum Go version to v1.22.

[^1]: https://www.freedesktop.org/wiki/Specifications/StatusNotifierItem/StatusNotifierItem/
[^2]: https://doc.qt.io/qt-6/qpixmap.html